gate para permissões do administrador

+++ app/Http/Controllers/Admin/GenreController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Genre;
+use Auth;
 
 class GenreController extends Controller
 {
@@ -15,6 +16,8 @@ class GenreController extends Controller
 
     public function index()
     {
+        $this->authorize('admin_operations');
+
         $genres = Genre::all();
         return view('admin.genre.index',[
             'genres' => $genres
@@ -27,6 +30,8 @@ class GenreController extends Controller
 
     public function create()
     {
+        $this->authorize('admin_operations');
+
         return view('admin.genre.create');
     }
 
@@ -36,6 +41,8 @@ class GenreController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('admin_operations');
+        
         $genre = new Genre;
         $genre->genre = $request->genre;
 
@@ -50,6 +57,8 @@ class GenreController extends Controller
 
     public function show($id)
     {
+        $this->authorize('admin_operations');
+
         $genre = Genre::findOrFail($id);
         
         return view('admin.genre.show', [
@@ -59,6 +68,8 @@ class GenreController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('admin_operations');
+
         $genre = Genre::findOrFail($id);
 
         return view('admin.genre.edit', [
@@ -68,6 +79,8 @@ class GenreController extends Controller
 
     public function update(Request $request, $id)
     {
+        $this->authorize('admin_operations');
+
         $genre = Genre::findOrFail($id);
         $genre->genre = $request->genre;
         $genre->save();
@@ -77,6 +90,8 @@ class GenreController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('admin_operations');
+
         Genre::findOrFail($id)->delete();
 
         return redirect('/admin/genre')->with('success_msg', "Gênero removido com sucesso");
